requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=conda_forge
_openmp_mutex=4.5=1_gnu
ca-certificates=2020.12.5=ha878542_0
certifi=2020.12.5=py38h578d9bd_1
cycler=0.10.0=py_2
decorator=4.4.2=py_0
freetype=2.10.4=h7ca028e_0
gpy=1.9.9=py38h8790de6_3
jpeg=9d=h36c2ea0_0
kiwisolver=1.3.1=py38h82cb98a_0
lcms2=2.11=hcbb858e_1
ld_impl_linux-64=2.35.1=hea4e1c9_1
libblas=3.9.0=6_openblas
libcblas=3.9.0=6_openblas
libffi=3.3=h58526e2_2
libgcc-ng=9.3.0=h5dbcf3e_17
libgfortran-ng=9.3.0=he4bcb1c_17
libgfortran5=9.3.0=he4bcb1c_17
libgomp=9.3.0=h5dbcf3e_17
liblapack=3.9.0=6_openblas
libllvm10=10.0.1=he513fc3_3
libopenblas=0.3.12=pthreads_h4812303_1
libpng=1.6.37=h21135ba_2
libstdcxx-ng=9.3.0=h2ae2ef3_17
libtiff=4.2.0=hdc55705_0
libwebp-base=1.1.0=h36c2ea0_3
llvmlite=0.35.0=py38h4630a5e_0
lz4-c=1.9.3=h9c3ff4c_0
matplotlib-base=3.3.3=py38h5c7f4ab_0
ncurses=6.2=h58526e2_4
numba=0.52.0=py38h51da96c_0
numpy=1.19.4=py38h18fd61f_2
olefile=0.46=pyh9f0ad1d_1
openssl=1.1.1i=h7f98852_0
pandas=1.2.0=py38h51da96c_0
paramz=0.9.5=py_0
pillow=8.1.0=py38h357d4e7_0
pip=20.3.3=pyhd8ed1ab_0
pydoe=0.3.8=py_1
pyparsing=2.4.7=pyh9f0ad1d_0
python=3.8.6=hffdb5ce_4_cpython
python-dateutil=2.8.1=py_0
python_abi=3.8=1_cp38
pytz=2020.5=pyhd8ed1ab_0
readline=8.0=he28a2e2_2
scipy=1.6.0=py38hb2138dd_0
setuptools=49.6.0=py38h924ce5b_2
simpy=4.0.1=pypi_0
six=1.15.0=pyh9f0ad1d_0
sqlite=3.34.0=h74cdb3f_0
tk=8.6.10=h21135ba_1
tornado=6.1=py38h25fe258_0
tzdata=2020f=he74cb21_0
wheel=0.36.2=pyhd3deb0d_0
xz=5.2.5=h516909a_1
zlib=1.2.11=h516909a_1010
zstd=1.4.8=ha95c52a_1
